requirements.txt
django
djangorestframework
djoser
psycopg2
djangorestframework-simplejwt